refactor entry points
